refactor: TransformControlPanel 삭제 및 TransformScreen 으로 UI 통합 (#150)
- `TransformControlPanel.kt` 파일을 삭제하고 해당 UI 컴포포넌트들을 `TransformScreen.kt` 내 `TransformContent`로 통합
- Preview 구성 시 `ZCardTheme`를 사용하도록 변경하여 테마 일관성 유지
